backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
alembic==1.7.1
smartapi-python==1.0.0
pandas==1.3.3
numpy==1.21.2
websockets==10.0
redis==4.0.1
psycopg2-binary==2.9.1
